requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
dash
scikit-learn
datetime
dash-bootstrap-components
dash-extensions
shap
umap-learn
streamlit